Removed `DefaultObjectFactory` singleton.
